+++ blas/coma_dsdot.h
@@ -6,6 +6,6 @@
 #define COMABLAS_COMA_DSDOT_H
 
 double sdsdot(unsigned int n, float b, float* x, int incx, float* y, int incy);
-double dsdot(unsigned int n, double * x, int incx, double * y, int incy);
+long double dsdot(unsigned int n, double * x, int incx, double * y, int incy);
 
 #endif //COMABLAS_COMA_DSDOT_H
